+++ tests/unit/test_add_group_and_access_to_model_yml.py
@@ -3,7 +3,7 @@ from dbt.contracts.graph.nodes import Group
 from dbt.contracts.graph.unparsed import Owner
 from dbt.node_types import AccessType, NodeType
 
-from dbt_meshify.storage.yaml_editors import DbtMeshYmlEditor
+from dbt_meshify.storage.file_content_editors import DbtMeshYmlEditor
 from tests.unit import read_yml
 
 meshify = DbtMeshYmlEditor()
